# ActiveModel::Dirty: only reset original values for persisted records